Refatoração do código para fins de legibilidade e manutenibilidade
